♻️ create tickets utils
