+++ packages/core-ethereum/src/channel/ticket.ts
@@ -17,10 +17,6 @@ class TicketStatic {
 
   constructor(public coreConnector: HoprEthereum) {}
 
-  // check before received ACK
-  // check after received ACK
-  // check when submitting
-
   public async submit(
     ticket: AcknowledgedTicket,
     ticketIndex: Uint8Array
@@ -45,11 +41,6 @@ class TicketStatic {
 
       const { hoprChannels, signTransaction, account, utils } = this.coreConnector
       const { r, s, v } = utils.getSignatureParameters((await ticket.signedTicket).signature)
-
-      // on-chain channel has funds
-      // all valid tickets amount should be 0.001
-      // on-chain "call" method to verify
-      // remove old UNACK tickets
 
       const hasPreImage = !u8aEquals(ticket.preImage, EMPTY_PRE_IMAGE)
       if (!hasPreImage) {
